project/requirements.txt
flask==3.0.2
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
pymysql==1.1.0

tensorflow==2.17.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0

